Modify intermediate code
Item and ItemState is modify to better indicate their uses. OriginData
is moved to ItemState, which is then passed between items.

`closure_parent` is now consistently set on all nodes, including peek
nodes, and is the primary means to produce follow items in the
`scan_items` function.

A pass is made to remove unused code and simplify the function
interfaces to better indicate their uses.

This also makes a correction to the binning of nodes based on their
origin Symbols within `handle_unresolved_scanner_nodes`.