refactor: reorganize runtime code, add test package, moves tests
Created a test package to handle end-to-end testing

Re-organize runtime package to better distinguish parser variants and
their component types

Adds forking parser. This takes the core parsing code of the error
recovery parser and adapts it for generalized parsers that can handle
ambiguous grammars. The error recovery parser will be rebuilt on top
of the forking parser.

Moves the `fork` rule from the `transitive_statement` production
to the `goto_sequence` production in ir.sg